# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vaya que valla",
        "version": "0.1"
    },
    "paths": {
...
        "/api/billboard_faces": {
            "get": {
                "tags": [
                    "Billboard_faces"
                ],
                "summary": "List billboard_faces with pagination",
                "operationId": "cf9f4dc1715a08b27e7c3d5c944d2251",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "itemsPerPage",
                        "in": "query",
                        "description": "Items per page",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Sort by field",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Billboard_faces"
                ],
                "summary": "Register billboardface",
                "operationId": "27892523bfe19f4de4e568410b36eadf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "billboard_id",
                                    "face",
                                    "location_detail"
                                ],
                                "properties": {
                                    "billboard_id": {
                                        "type": "number",
                                        "maxLength": 20
                                    },
                                    "face": {
                                        "type": "string",
                                        "maxLength": 10
                                    },
                                    "location_detail": {
                                        "type": "string",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "BillboardFace registered successfully"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/billboard_faces/{id}": {
            "get": {
                "tags": [
                    "Billboard_faces"
                ],
                "summary": "Get billboardface by ID",
                "operationId": "512f48302b91e263b9721c4bfadd333d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the billboardface",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "BillboardFace found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "BillboardFace found"
                                        },
                                        "organization": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "BillboardFace not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Billboard_faces"
                ],
                "summary": "Update billboardface",
                "operationId": "e1b62c89e97f35b0d9945465b690ec2a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the billboardface to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated billboardface data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "billboard_id",
                                    "face",
                                    "location_detail"
                                ],
                                "properties": {
                                    "billboard_id": {
                                        "type": "number",
                                        "maxLength": 20
                                    },
                                    "face": {
                                        "type": "string",
                                        "maxLength": 10
                                    },
                                    "location_detail": {
                                        "type": "string",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "BillboardFace updated successfully"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token JWT con el formato 'Bearer {token}'.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}